adding abstract intensifier class
